+++ docs/changelog.rst
@@ -5,6 +5,11 @@ Changelog
 ------
 *DD MMM YYYY*
 
+**Tools**
+
+* Fixed opening projects from the menubar when in the welcome panel.
+* Fixed projects opened from the menubar not being added to the recent projects list.
+
 0.44.0
 ------
 *13 May 2021*